requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
cerebras_cloud_sdk>=1.0.0
markdown>=3.4.0
